gymnasium/requirements.txt
gymnasium==1.2.0
numpy==2.3.1
pyzmq==27.0.0
